Debug truss_stress.jl coupled variable R (#152)
* minor correction

the argument in visualize does not have "u", so I could not show the Makie figure. 

I recheck with the VDO by Yijiang https://www.youtube.com/watch?v=sBqdkxPXluU
and corrected them as in the clip, and now it works fine.

* enable visualization

* Update JSON description

* TrussStress debug

ndim = 3

* minor correction

the argument in visualize does not have "u", so I could not show the Makie figure. 

I recheck with the VDO by Yijiang https://www.youtube.com/watch?v=sBqdkxPXluU
and corrected them as in the clip, and now it works fine.

* enable visualization

* Remove Visualization related lines

* Remove Visualization related lines

* Remove Visualization related lines

* Uncomment the module

Uncomment line 1 module that was accidentally commented.

* Update truss_stress.jl

* Debug TrussStress

add copy(R) on  line 35

* Update truss_stress.jl

* Update truss_compliance_2d1.jl

* Update truss_compliance_2d2.jl

* Update truss_compliance_3d2.jl

* add test file for truss_stress.jl

* add testset

* Try update a commit

* organize truss_stress test files

* Update docs for truss JSON format

---------

Co-authored-by: Mohamed Tarek <mohamed82008@gmail.com>
Co-authored-by: pitipat_acer <miraithn@gmail.com>
Co-authored-by: yijiangh <huang@hey.com>
Co-authored-by: Yijiang Huang <yijiang94817@gmail.com>