Add time stepping to hyperelastic solver
Additional validation is pending regarding the time stepping code, but confidence is fairly high that it is implemented correctly. Stability seems to be excellent so long as sufficiently small time steps are used to achieve the solution. This also includes some of the foundation for implenting a nearly-incompressible hyperelastic solver that is not implemented yet.